ui_embeddings/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "ui_embeddings"
version = "0.1.0"
dependencies = [
    "streamlit>=1.37.0",
    "openai==2.2.0",
    "weaviate-client==4.17.0",
    "boto3==1.40.40",
    "boto3-stubs[s3]==1.40.40",
]

[tool.setuptools]
packages = ["ui_embeddings"]

[project.optional-dependencies]
dev = [
    "black==25.9.0",
    "isort==6.0.1",
    "build==1.3.0",
    "pip-tools==7.5.0",
    "mypy",
]
